requirements.txt
asgiref==3.8.1
Django==5.1.6
django-js-asset==3.1.2
django-prose-editor==0.10.4
djangorestframework==3.15.2
nh3==0.2.21
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
